+++ DataLayers/Layers/AccessibleLayer.cs
@@ -34,7 +34,7 @@ namespace Pathoschild.Stardew.DataLayers.Layers
 
         /// <summary>The action tile property values which trigger a warp.</summary>
         /// <remarks>See remarks on <see cref="IsWarp"/>.</remarks>
-        private readonly HashSet<string> WarpActions = new HashSet<string> { "EnterSewer", "LockedDoorWarp", "Warp", "WarpCommunityCenter", "WarpGreenhouse", "WarpMensLocker", "WarpWomensLocker", "WizardHatch" };
+        private readonly HashSet<string> WarpActions = new HashSet<string> { "EnterSewer", "LockedDoorWarp", "Mine", "Warp", "WarpCommunityCenter", "WarpGreenhouse", "WarpMensLocker", "WarpWomensLocker", "WizardHatch" };
 
         /// <summary>The touch action tile property values which trigger a warp.</summary>
         private readonly HashSet<string> TouchWarpActions = new HashSet<string> { "Door", "MagicWarp" };
@@ -89,6 +89,9 @@ namespace Pathoschild.Stardew.DataLayers.Layers
             {
                 foreach (Building building in buildableLocation.buildings)
                 {
+                    if (building.indoors.Value == null || (building.humanDoor.X < 0 && building.humanDoor.Y < 0))
+                        continue;
+
                     buildingDoors.Add(new Vector2(building.humanDoor.X + building.tileX.Value, building.humanDoor.Y + building.tileY.Value));
                     buildingDoors.Add(new Vector2(building.humanDoor.X + building.tileX.Value, building.humanDoor.Y + building.tileY.Value - 1));
                 }
@@ -138,6 +141,23 @@ namespace Pathoschild.Stardew.DataLayers.Layers
             // check tile touch actions
             Tile backTile = location.map.GetLayer("Back").PickTile(new Location(tilePixels.X, tilePixels.Y), Game1.viewport.Size);
             if (backTile != null && backTile.Properties.TryGetValue("TouchAction", out PropertyValue touchAction) && this.TouchWarpActions.Contains(touchAction.ToString().Split(' ')[0]))
+                return true;
+
+            // check map warps
+            try
+            {
+                if (location.isCollidingWithWarpOrDoor(tilePixels) != null)
+                    return true;
+            }
+            catch
+            {
+                // This fails in some cases like TMX Loader's custom tile properties. It's safe to
+                // ignore the error here, since that means it's not a valid warp.
+            }
+
+            // check mine ladders/shafts
+            const int ladderID = 173, shaftID = 174;
+            if (location is MineShaft && buildingTile != null && (buildingTile.TileIndex == ladderID || buildingTile.TileIndex == shaftID) && buildingTile.TileSheet.Id == "mine")
                 return true;
 
             return false;